Moved delegate creator into ReflectionExtensions and refactored it to alllow compilation of the AddHandler delegate to be done during the construction of an AttachedEvent instead of when the object it set during the ObjectBuilder phase.
