+++ Glass.Core/ReflectionExtensions.cs
@@ -3,19 +3,17 @@
     using System;
     using System.Collections.Generic;
     using System.Linq.Expressions;
+    using System.Reflection;
 
     public static class ReflectionExtensions
     {
         public static string GetFullPropertyName<T, TProperty>(this Expression<Func<T, TProperty>> propertySelector)
-
         {
             MemberExpression memberExp;
 
             if (!TryFindMemberExpression(propertySelector.Body, out memberExp))
-
                 return string.Empty;
-
-
+            
             var memberNames = new Stack<string>();
 
             do
@@ -30,12 +28,10 @@
         // code adjusted to prevent horizontal overflow
 
         private static bool TryFindMemberExpression(this Expression exp, out MemberExpression memberExp)
-
         {
             memberExp = exp as MemberExpression;
 
             if (memberExp != null)
-
                 return true;
 
 
@@ -47,7 +43,6 @@
             // ...which are the cases checked in IsConversion
 
             if (IsConversion(exp) && exp is UnaryExpression)
-
             {
                 memberExp = ((UnaryExpression) exp).Operand as MemberExpression;
 
@@ -61,9 +56,34 @@
         }
 
         private static bool IsConversion(Expression exp)
-
         {
             return (exp.NodeType == ExpressionType.Convert) || (exp.NodeType == ExpressionType.ConvertChecked);
+        }
+
+
+
+        public static Func<object, Delegate> GetDelegateWithDefaultParameterValuesBound(this MethodInfo method)
+        {
+            var instanceParameter = Expression.Parameter(typeof(object));
+            var methodParameters = method.GetParameters();
+            var parameterExpressions = new Dictionary<int, ParameterExpression>();
+            var arguments = new List<Expression>(methodParameters.Length);
+            for (int i = 0; i < methodParameters.Length; i++)
+            {
+                if (methodParameters[i].HasDefaultValue)
+                {
+                    arguments.Add(Expression.Constant(methodParameters[i].DefaultValue));
+                }
+                else
+                {
+                    parameterExpressions.Add(i, Expression.Parameter(methodParameters[i].ParameterType));
+                    arguments.Add(parameterExpressions[i]);
+                }
+            }
+
+            return Expression.Lambda<Func<object, Delegate>>(Expression.Lambda(
+                    Expression.Call(Expression.Convert(instanceParameter, method.DeclaringType), method, arguments),
+                parameterExpressions.Values), instanceParameter).Compile();
         }
     }
 }
