+++ Glass.Core/Utils.cs
@@ -1,12 +1,9 @@
 ﻿namespace Glass.Core
 {
-    using System;
-    using System.Collections.Generic;
     using System.Linq;
-    using System.Linq.Expressions;
     using System.Reflection;
 
-    public static class Utils
+    public class Utils
     {
         public static void UniversalAdd(object collection, object item)
         {
@@ -25,27 +22,6 @@
             {
                 // handle your error
             }
-        }
-
-        public static Delegate GetDelegateWithDefaultParameterValuesBound(object instance, MethodInfo method)
-        {
-            var methodParameters = method.GetParameters();
-            var parameterExpressions = new Dictionary<int, ParameterExpression>();
-            var arguments = new List<Expression>(methodParameters.Length);
-            for (int i = 0; i < methodParameters.Length; i++)
-            {
-                if (methodParameters[i].HasDefaultValue)
-                {
-                    arguments.Add(Expression.Constant(methodParameters[i].DefaultValue));
-                }
-                else
-                {
-                    parameterExpressions.Add(i, Expression.Parameter(methodParameters[i].ParameterType));
-                    arguments.Add(parameterExpressions[i]);
-                }
-            }
-
-            return Expression.Lambda(Expression.Call(Expression.Constant(instance), method, arguments), parameterExpressions.Values).Compile();
         }
     }
 }
